refactor: localize audio and badge box initialization
